Update demo app to use nightly TFLite build instead of latest release build.
When the demo app updates to use a backwards-incompatible change to the TFLite Java API at HEAD, it'll fail to build on the old release (which is missing the API change).

Using the nightly build means the demo app will use a relatively fresh TFLite
build with API changes, in addition to other improvements. The user may need to
pull the latest demo code to keep up.

PiperOrigin-RevId: 200672004