Fix usage of `docs_controls.should_skip`.
`docs_controls` was recently added to allow users to tag objects with `do_not_generate_docs` (it can be used as a decorator). This allows you to skip objects by _identity_, instead of path.

In my initial implementation I checked for the tag in the doc generation step.

This change fixes it to skip objects during the api-crawling step. This is necessary to avoid crawling the internals of all the excluded objects.

To enable this, the `_is_private` method of PublicAPIVisitor needs access to the object, not just the path to it. The changes in `public_api` allow this.

`generate_lib` uses the `_is_private` signature change to inspect the object.

PiperOrigin-RevId: 209587765