Updates
Thank you for starting this! Here's a round of edits. I think it's almost ready to go. Could you take a look and see if there's anything we can improve in this version? When you're happy with it, please submit a PR to the TF Docs repo, and we can continue refining there. I'd like to add some graphics of the embedding projector before we publish, add more references to educational resources, and improve the intro to embeddings before we publish, but we can work on those changes in the docs PR whenever you're ready. Thanks again!