Update docker.md
The statement "`nvidia-docker` is not necessary" is incorrect. The `nvidia-docker` and `nvidia-docker2` packages are deprecated because the Docker runtime naively supports NVIDIA GPUs. But Docker does not support Nvidia GPUs on its own. The `nvidia-container-runtime` is still required.

Updated all Docker commands to use the command line flags that nvidia-container-toolkit uses. All users with fresh installations will be using nvidia-container-toolkit instead of `nvidia-docker` v2.