multi_worker_with_estimator.ipynb: Remove softmax to get logits
The rest of the `model_fn` assumes that the output from the model is logits,
so remove the softmax activation function from the last layer. Training did
not converge in my testing because the loss calculation assumed the inputs
were logits (`from_logits=True`) when they in fact were the output of a softmax.