Fix decorator handling in tensorflow api_generator.
The generator fails if you don't unwrap TFDecorators. Subproject functions decorated with `@tf.function` currently fail at `getfullargspec`, this is the way to unwrap them.

Also capture the tf_decorator names and add them to the docs.

Also removes the monkey-patching in generate2.py, this just does the unwraping automatically.

This adds 200 of missing decorators to the signature blocks of functions in tensorflow. But there are a lot of internal decorators, so I'm white-listing ones that I think are informative to users. There are no other changes to the output.

PiperOrigin-RevId: 291218898