Edit string shortcut for model compilation
The document states that string identifiers can be used as shortcut, yet the example given is no different from the compilation conducted above (presumably without strings identifiers, which is still confusing given that metrics was defined with a string identifier). The edit adds clarity by demonstrating how models can be compiled with and without string identifiers.