Fixed all the broken links by appending .md
The links did not have a `.md` at the end for some reason. This commit basically replaces `"><code>` with `.md"><code>` so that all the links will now map properly instead.

I am not sure why any of the links did not a .md at the attend. I tried with all the links, none of the links worked before, only working after I appended a .md at the URL's place to the original URL string.