nblint: Capture stderr messages within lint functions and pretty-print with status. Add basic parsing of command-line args.
Within a @lint function, any writes to stderr will be captured and presented as part of a more readable status summary at the end of all runs. This will make it easier to present useful errors/instructions within the CI interface. Writes to `sys.stdout` work as normal, printing when the function is run, instead of logging at the end of all runs.

PiperOrigin-RevId: 317336622