Update overfit_and_underfit.ipynb
"Underfitting occurs when there is still room for improvement on the test data." should read as "Underfitting occurs when there is still room for improvement on the train data.". Reason being the underfitting fails to extract adequate structure from train data and the "room for improvement" is dependent on train data and not test data!