Update data.ipynb
The modified section had a defect. The only reason the defect was not visible in the output was because the label width of 5 is exactly half of the feature length of 10. In that case the index ``[:-label_length]`` will be same as ``[:label_length]``. The correct index is ``[:label_length]``.

I have changed the label length to 3 so that any problem with the code shows up.

On a related note, I see that a label length of 5 is used throughout this notebook. I do not have the time right now to see if similar defect exists elsewhere.