Make the plot_attention function more robust.
for PR#1861

With an odd number of output word, the last element falls to the third line.
When 5,3 or 1 words are generated there is no third line.

This change uses `ceil` so that when the number of words is odd there's an empty cell on the last row. 5 words uses a 3x3 grid, 3 words is 2x2. The minimum grid size is 2x2.

PiperOrigin-RevId: 366155062