Just fixed a small typo ("Shakespear") 😅
Hi. I didn't push this change just to be a contributor but noticed this small typo, so thought to change it. I don't even know if it's a typo or an intended one. Thanks a lot for TensorFlow, TF had helped me a lot in my career and my academics ❤️