Bugfix in concatenating gradient batches
tf.stack only accounted for the last batch of gradients, instead of collecting the gradients from all batches. The all but the last batch were discarded. The correct implementation is to concatenate the gradient batches by the first axis. This way, all the gradients are correctly accounted for.

Tested the changes by running it in googe colab.