Add extra `set_random_seed` calls to `model_mapping.ipynb`.
Under the nightly TF package, this section of the notebook is failing as the default kernel initializers are not picked up by `DeterministicRandomTestTool`.

This could be fixed by passing `kernel_initializer=tf.compat.v1.random_normal_initializer()` to each layer, but I think that distracts too much from the migration code. So while this is a bit of a hack, it's for clarity's sake, and should be clear that it's not something a user needs to do.

Original breakage caused by [this commit](https://github.com/keras-team/keras/commit/69b41cf4282682f9443dc65528da5fe487cd104f).

PiperOrigin-RevId: 468603054