Move TF1 hashing example out of code cell.
A planned change in TensorFlow will disable hashing
of tensors in eager execution mode even when
numpy-style equality is disabled
(the TensorFlow 1 default).

This will break the moved example. Since we cannot
disable and re-enable eager execution after the
beginning of the program, this example is no longer
runnable under TF2 - even with the compatibility
layer.

As noted in the edited documentation, `set` requires
its elements to be hashable - since this example runs
under TF2 and we will no longer allow hashing of tensors
in eager execution mode full-stop, the code _is_ valid
TF1 code but cannot be run with the proposed changes to TF2.

PiperOrigin-RevId: 520100823