To fix `nbformat` pre-commit instructions...
The current instructions did not work from me and didn't seem to match the usage in the [pre-commit docs](https://pre-commit.com/).

I believe that these changes (almost) address the issue by

a) specifying that the YAML contents go in a `.pre-commit-config.yaml` file`;
b) clarifying that `.pre-commit-hooks.yaml` is a pre-commit plugin; and
c) completing the config YAML by specifying which `hooks` to run (from the plugin)

However, this still won't work because `.pre-commit-hooks.yaml` is not located at the root of the repo (as required in https://pre-commit.com/#new-hooks). Am I missing something? E.g. is tools/tensorflow_docs/tools intended as some sort of sub-repo?